[REFACTOR] 기사 키워드에 원본 기사 Id를 멤버변수로 포함
