+++ src/main/java/com/example/whiplash/user/service/UserKeywordService.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.example.whiplash.apiPayload.ErrorStatus;
 import com.example.whiplash.apiPayload.exception.WhiplashException;
 import com.example.whiplash.user.domain.User;
-import com.example.whiplash.user.domain.keyword.Keyword;
-import com.example.whiplash.user.domain.keyword.UserKeyword;
+import com.example.whiplash.keyword.user.Keyword;
+import com.example.whiplash.keyword.user.UserKeyword;
 import com.example.whiplash.user.repository.keyword.KeywordRepository;
 import com.example.whiplash.user.repository.keyword.UserKeywordRepository;
 import com.example.whiplash.user.repository.user.UserRepository;
