+++ src/main/java/com/example/whiplash/user/service/UserService.java
@@ -4,8 +4,8 @@ package com.example.whiplash.user.service;
 import com.example.whiplash.apiPayload.ErrorStatus;
 import com.example.whiplash.apiPayload.exception.WhiplashException;
 import com.example.whiplash.converter.InvestorProfileConverter;
-import com.example.whiplash.user.domain.keyword.Keyword;
-import com.example.whiplash.user.domain.keyword.UserKeyword;
+import com.example.whiplash.keyword.user.Keyword;
+import com.example.whiplash.keyword.user.UserKeyword;
 import com.example.whiplash.user.domain.profile.InvestorProfile;
 import com.example.whiplash.user.domain.Role;
 import com.example.whiplash.user.domain.User;
